requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
everapi==0.1.1
freecurrencyapi==0.1.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
